fix: DRY
